refactor: update dropdown import
